[libutee] add function: dump_active_sources
This functions dumps the list of all IPs it is balancing into a file
that can be specified.

If the output file is named "${dirname}/${fname}", utee writes the list
into a file named "${dirname}/.${fname}.tmp". After the write is
completed and the file closed, utee renames it to "${dirname}/${fname}".
This ensures that any process reading from "${dirname}/${fname}" will
always read a fully written file.

utee not only outputs the currently sending IPs, but also IPs it has
seen in the past together with how many iterations/dumps it has seen
them last.

The final output file looks like:

IP3<tab>50
IP1<tab>0
IP2<tab>1

Here,

IP1 has seen data in the since the last write,
IP2 did not send any data during the last rotation, but did the one
before. So it may only be slow, have a hiccup or simply did not
throughput enough data
IP3 did not see any data for a while, so it is either broken or
de-commisioned.

Be aware: utee does not sort the content of the file in any way. The
order is purely random (actually... it depends on how the hash table is
being traversed).

Finally, IPs will be removed once their counter goes over a certain
(configureable) threshold.

Signed-off-by: Robert Wuttke <robert@benocs.com>